+++ CHANGES.rst
@@ -1,6 +1,12 @@
 Changes
 =======
 
+0.1.3 -- 2016-mm-dd
+-------------------
+* Removed deprecated methods "eps", "svg", "png", "pdf", and "txt" from
+  segno.QRCode
+
+
 0.1.2 -- 2016-08-17
 -------------------
 * Updated docs